Rename cc files to cpp and add license
