--- tools/protoc-gen-chpl/protoc-gen-chpl.cc
@@ -1,8 +1,0 @@
-#include <google/protobuf/compiler/plugin.h>
-
-#include <generator.hh>
-
-int main(int argc, char *argv[]) {
-  chapel::Generator generator;
-  return PluginMain(argc, argv, &generator);
-}
