+++ tools/protoc-gen-chpl/repeated_enum_field.cpp
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2020 Hewlett Packard Enterprise Development LP
+ * Copyright 2004-2019 Cray Inc.
+ * Other additional copyright holders may be indicated within.
+ *
+ * The entirety of this work is licensed under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include <repeated_enum_field.hh>
+
+namespace chapel {
+
+  RepeatedEnumFieldGenerator::RepeatedEnumFieldGenerator(
+      const FieldDescriptor* descriptor)
+      : FieldGeneratorBase(descriptor) {
+  }
+
+  RepeatedEnumFieldGenerator::~RepeatedEnumFieldGenerator() {
+
+  }
+
+  void RepeatedEnumFieldGenerator::GenerateMembers(Printer* printer) {
+    printer->Print(
+      variables_,
+      "var $property_name$: list($type_name$);\n"
+      "proc $name$ { return $property_name$; }\n"
+      "proc ref $name$ ref { return $property_name$; }\n"
+    );
+  }
+
+}  // namespace chapel
